requirements.txt
# runtime
pandas>=2.2,<3
numpy>=1.26,<3
plotly>=5.22,<6
streamlit>=1.36,<2
python-dotenv>=1.0,<2
openai>=1.40,<2
tiktoken>=0.7,<1
scikit-learn>=1.5,<2
umap-learn>=0.5,<1
scipy>=1.13,<2
kaleido>=0.2,<1
statsmodels>=0.14
pmdarima>=2.0      # optional
prophet==1.1.5 ; python_version < "3.12"
cmdstanpy>=1.2 ; python_version < "3.12"